Use vsvarsall.bat to get environment values, instead of user prefs.
If it is possible to deduce the locations of VSINSTALLDIR and
VCINSTALLDIR, then run the vsvarsall.bat file in a subprocess with a
stripped down environment.

Get the values of LIB, INCLUDE, PATH, and LIBPATH from that
subprocess' environment.

This means we don't fish around in the registry and try to read things
from the user's XML settings file. The XML settings are a lot to
parse, and using them means a change in the IDE settings could affect
batch builds, which would be bad. It also saves a lot of time.

It also means that the SCons environment setting MSVS8_PLATFORM can
determine which platform we build for, as it is passed to the
vsvarsall.bat file. So if we were to set MSVS8_PLATFORM='x86_amd64' in
an Environment() constructor, we'd be compiling with the 64-bit
tools. Thus, a lot of mess could be cleaned out of the DevTests'
SConstruct.

And we'd be using the environment settings that makefiles and batch
scripts end up using.

Copied from: svn+ssh://svn.dlogics.com/svn/dev/Packages/scons-local/trunk@31940